backend/requirements.txt
Flask==2.0.1
Flask-CORS==3.0.10
requests==2.26.0
python-dotenv==0.19.1
Pillow==10.4.0
markdown==3.5.1
pytest==6.2.5




